--- src/pages/catalog-page/catalog/filtering/options/category-options/enums.ts
@@ -1,4 +1,0 @@
-export enum DEFAULT_VALUE {
-  CATEGORY = 'default-category',
-  SUBCATEGORY = 'default-subcategory',
-}
