+++ src/pages/catalog-page/catalog/filtering/options/category-options/handle-category-change.ts
@@ -2,7 +2,7 @@ import type { NavigateFunction } from 'react-router-dom';
 import { catalogStore } from '../../../../../../store/catalog-store';
 import { PagePath } from '../../../../../../router/enums';
 import { generatePath } from 'react-router-dom';
-import { DEFAULT_VALUE } from './enums';
+import { DEFAULT_VALUE } from '../../../../../../sources/enums/default-values';
 
 export const handleCategoryChange = (
   event: React.ChangeEvent<HTMLSelectElement>,
